fix clang-tidy issue according to clang-tidy 14.

+++ arena/align.hpp
@@ -37,7 +37,7 @@ template <uint64_t N>
 
 [[gnu::always_inline]] inline auto AlignUp(uint64_t n, uint64_t block_size) noexcept -> uint64_t {
     uint64_t reminder = n % block_size;
-    return n - reminder + (static_cast<int>(static_cast<bool>(reminder))) * block_size;
+    return n - reminder + (static_cast<uint64_t>(static_cast<bool>(reminder))) * block_size;
 }
 
 }  // namespace stdb::memory::align
